Introducing Process.enso_builder
