+++ distribution/lib/Standard/Base/0.0.0-dev/src/System/Process.enso
@@ -5,6 +5,7 @@ import project.System.Process.Process_Builder.Process_Result
 from project.Data.Boolean import Boolean, False
 from project.Metadata.Widget import Text_Input, Vector_Editor
 
+polyglot java import org.enso.base.file_system.File_Utils
 ## ---
    aliases: [run a command]
    group: System
@@ -44,7 +45,6 @@ run command:Text arguments=[] stdin:Text="" redirect_out_err:Boolean=False =
     - `command`: The command to execute on the system.
     - `arguments`: The arguments to pass to `command`. These must be text.
     - `stdin`: Any content to pass to the standard input for `command`.
-    - `redirect_out_err`: Redirect stdout/stderr to `Process_Result`.
 
    ## Examples
    ### Create a new builder for a command "echo".
@@ -56,3 +56,11 @@ run command:Text arguments=[] stdin:Text="" redirect_out_err:Boolean=False =
    ```
 new_builder : Text -> Vector Text -> Text -> Process_Builder
 new_builder command arguments=[] stdin="" = Process_Builder.Value command arguments stdin
+
+## ---
+   private: true
+   advanced: true
+   ---
+   Create a new Enso process builder.
+
+enso_builder arguments=[] stdin="" = Process_Builder.Value File_Utils.getExecutableLocation arguments stdin
