+++ test/Base_Tests/src/Enso_Bin_Utils.enso
@@ -7,7 +7,6 @@ polyglot java import java.lang.Process as J_Process
 polyglot java import java.lang.String as J_String
 polyglot java import java.lang.System as J_System
 polyglot java import java.io.File as J_File
-polyglot java import org.enso.base_test_helpers.ExecutableLocation as J_ExecutableLocation
 
 
 ## Returns True if current process runs in AOT mode (Native Image).
@@ -27,19 +26,16 @@ create_tmp_dir -> File =
 
 ## Absolute path to the Enso binary.
 enso_bin -> Text =
-    J_ExecutableLocation.getExecutableLocation
+    Error.throw "Cannot get enso_bin path right now"
 
 
 ## Runs an Enso subprocess.
 enso_run (args : Vector Text) (cwd : Text = File.current_directory.path) -> Text =
-    proc_bldr = J_ProcessBuilder.new [enso_bin]+args
-    proc_bldr.redirectErrorStream True
-    proc_bldr.directory (J_File.new cwd)
-    proc = proc_bldr.start
-    retcode = proc.waitFor
-    bytes = proc.getInputStream.readAllBytes
-    out = J_String.new bytes
-    if retcode != 0 then
+    proc_bldr = Process.enso_builder args
+    # proc_bldr.directory (J_File.new cwd)
+    proc = proc_bldr.create redirect_out_err=True
+    out = proc.stdout
+    if proc.exit_code != 0 then
         Test.fail out
     out
 
@@ -48,7 +44,7 @@ create_new_project parent_dir:File proj_name:Text main_src:Text -> File =
     Runtime.assert parent_dir.is_directory
     proj_dir = parent_dir / proj_name
     args = ["--new", proj_dir.path, "--new-project-name", proj_name]
-    ret = Process.run enso_bin args
+    ret = Process.enso_builder args . create
     ret.exit_code . should_equal Exit_Code.Success
     main = proj_dir / "src" / "Main.enso"
     Runtime.assert main.exists
